==> 02_check_smart_url.ipynb <==
import pandas as pd
pd.options.display.float_format = '{:.2f}'.format # scientific notation
import csv
from ast import literal_eval
import math
import time

url = pd.read_csv("data/smarturl20221005.csv").drop(["est_total"], axis=1)

url.head()

# create new column: 1- yes smart city keywords, 0- no smart city keywords
url["smart"] = np.where(url.total==0, 0, 1)

# number of cities with smart city keywords
url.smart.sum()

# cities with smart city keywords
smart = url[url.smart==1]

smart.head()

# save to csv for manual checking
smart.to_csv("smart_manual.csv", index=False)
