==> 03_merge_checked_df.ipynb <==
import pandas as pd
pd.options.display.float_format = '{:.2f}'.format
import csv
from ast import literal_eval
import math
import time

# import data
df = pd.read_csv("data/all_url20220929.csv") # all the cities
checked = pd.read_csv("data/smart_manual_TD_MC_checked.csv") # manually checked list of smart cities
loc = pd.read_csv("data/nhgis/nhgis0009_ds248_2020_place.csv") # can be downloaded from https://www.nhgis.org/

# keep useful columns
df = df[["GISJOIN", "POP20", "BASENAME", "State"]]
loc = loc[["GISJOIN", "INTPTLAT", "INTPTLON"]]
checked = checked[["GISJOIN", "manual_check"]]
# rename manual_check to smart, meaning yes a smart city
checked.rename(columns={"manual_check": "smart"}, inplace=True)

# merge into one df
df = df.merge(checked, on="GISJOIN", how="left")
df = df.merge(loc, on="GISJOIN", how="left")

# fill with 0 if no website
df.loc[df.smart!=1, "smart"] = 0

df.to_csv("data/smart_cities20240110.csv", index=False)
