==> 06_for_hpcc.ipynb <==
import csv
import pandas as pd

df = pd.read_csv("data/smart_manual_TD_MC_checked.csv")

df = df[df.manual_check==1] # filter out non-smart cities
df = df[["GISJOIN","link"]] # keep webpage links and GISJOIN

df

df.to_csv("data/for_hpcc.csv", index=False)
