==> 08_combine_text.ipynb <==
import pandas as pd
import csv
from ast import literal_eval
import sys
import numpy as np
import glob

sc = pd.read_csv("data/smart_cities20240110.csv")

sc = sc[sc.smart==1]

path = "../hpcc/text/all"
file_names = glob.glob(path+"/*.csv")

file_list = []
for file_name in file_names:
    file = pd.read_csv(file_name, names=["GISJOIN","text"])
    file_list.append(file)

df = pd.concat(file_list, ignore_index=True)

df = df.dropna(subset=["text"]) # drop empty text

df = df.loc[~df.text.str.contains("access denied", case=False)] # drop "access denied"

sorter = list(sc.GISJOIN)
df.GISJOIN = df.GISJOIN.astype("category")
df.GISJOIN = df.GISJOIN.cat.set_categories(sorter)

df

part1 = df[:2000]
part2 = df[2000:]

part1.to_csv("data/hpcc_text_part1_20240124.csv", index=False)
part2.to_csv("data/hpcc_text_part2_20240124.csv", index=False)
