==> 10_reformat_sdgs.ipynb <==
import pandas as pd
import csv
from ast import literal_eval
import sys
import numpy as np
import glob
import re

path = "data/seesus_text"
file_names = glob.glob(path+"/*.csv")

# combine all text files to one
file_list = []
for file_name in file_names:
    gisjoin = re.findall(r'[G]\d{8}', file_name)[0]
    file = pd.read_csv(file_name, names=["sdg","target","see"])
    file["GISJOIN"] = gisjoin
    file_list.append(file)

df = pd.concat(file_list, ignore_index=True)

df = df[["GISJOIN", "sdg", "see", "target"]] #reorder

# expand to columns
sdg1 = []
sdg2 = []
sdg3 = []
sdg4 = []
sdg5 = []
sdg6 = []
sdg7 = []
sdg8 = []
sdg9 = []
sdg10 = []
sdg11 = []
sdg12 = []
sdg13 = []
sdg14 = []
sdg15 = []
sdg16 = []
sdg17 = []
soc = []
env = []
eco = []

for i in range(len(df)):
    gisjoin = df.GISJOIN.iloc[i]
    if "'SDG1'" in df.sdg.iloc[i]:
        sdg1.append(1)
    else:
        sdg1.append(0)
    if "'SDG2'" in df.sdg.iloc[i]:
        sdg2.append(1)
    else:
        sdg2.append(0)
    if "'SDG3'" in df.sdg.iloc[i]:
        sdg3.append(1)
    else:
        sdg3.append(0)
    if "'SDG4'" in df.sdg.iloc[i]:
        sdg4.append(1)
    else:
        sdg4.append(0)
    if "'SDG5'" in df.sdg.iloc[i]:
        sdg5.append(1)
    else:
        sdg5.append(0)
    if "'SDG6'" in df.sdg.iloc[i]:
        sdg6.append(1)
    else:
        sdg6.append(0)
    if "'SDG7'" in df.sdg.iloc[i]:
        sdg7.append(1)
    else:
        sdg7.append(0)
    if "'SDG8'" in df.sdg.iloc[i]:
        sdg8.append(1)
    else:
        sdg8.append(0)
    if "'SDG9'" in df.sdg.iloc[i]:
        sdg9.append(1)
    else:
        sdg9.append(0)
    if "'SDG10'" in df.sdg.iloc[i]:
        sdg10.append(1)
    else:
        sdg10.append(0)
    if "'SDG11'" in df.sdg.iloc[i]:
        sdg11.append(1)
    else:
        sdg11.append(0)
    if "'SDG12'" in df.sdg.iloc[i]:
        sdg12.append(1)
    else:
        sdg12.append(0)
    if "'SDG13'" in df.sdg.iloc[i]:
        sdg13.append(1)
    else:
        sdg13.append(0)
    if "'SDG14'" in df.sdg.iloc[i]:
        sdg14.append(1)
    else:
        sdg14.append(0)
    if "'SDG15'" in df.sdg.iloc[i]:
        sdg15.append(1)
    else:
        sdg15.append(0)
    if "'SDG16'" in df.sdg.iloc[i]:
        sdg16.append(1)
    else:
        sdg16.append(0)
    if "'SDG17'" in df.sdg.iloc[i]:
        sdg17.append(1)
    else:
        sdg17.append(0)
    if "'social_sustainability': True" in df.see.iloc[i]:
        soc.append(1)
    else:
        soc.append(0)
    if "'environmental_sustainability': True" in df.see.iloc[i]:
        env.append(1)
    else:
        env.append(0)
    if "'economic_sustainability': True" in df.see.iloc[i]:
        eco.append(1)
    else:
        eco.append(0)

df["SDG1"]=sdg1
df["SDG2"]=sdg2
df["SDG3"]=sdg3
df["SDG4"]=sdg4
df["SDG5"]=sdg5
df["SDG6"]=sdg6
df["SDG7"]=sdg7
df["SDG8"]=sdg8
df["SDG9"]=sdg9
df["SDG10"]=sdg10
df["SDG11"]=sdg11
df["SDG12"]=sdg12
df["SDG13"]=sdg13
df["SDG14"]=sdg14
df["SDG15"]=sdg15
df["SDG16"]=sdg16
df["SDG17"]=sdg17
df["soc"]=soc
df["env"]=env
df["eco"]=eco

df = df.drop(['sdg','see'], axis=1)

df.to_csv("data/sdgs20240202.csv", index=False)
