==> 12_reformat_targets.ipynb <==
import pandas as pd
import csv
from ast import literal_eval
import sys
import numpy as np
import glob
import re

df = pd.read_csv("data/sdgs20240202.csv", converters={"target": literal_eval})

df = df[["GISJOIN", "target"]] #reorder

df.head(10)

#df.to_csv("data/targets20240202.csv", index=False)
